moving NotificationContainer into component folder
